Add a super "dumb" cache
Does what it says on the tin with a hash. Lookup time goes from 0.016s
down to 0.00071s on the second request. Boom!

The way this works is by adding a "match anything" block at the top that
bails out early if we don't have the URL in the cache. On the upstream
lookup side, when we get an answer back for a regular ol' URL, add it to
the cache. We could handle other types, like CNAME, MX, TXT, etc., but I
don't wanna mess with that yet.