Implement a shared DNS cache for Rails and dns_server
TIL a good bit about caching in Rails. Taking a note from how we setup
`Redis.instance` in PCO products, I did the same with `Redis.dns_cache`.
This allows both Rails and `dns_server` to read from the same source of
truth (which is way faster than relying on database reads). When
`dns_server` fires up, chuck all the DnsRecords into the cache so that
the first `match` block can read 'em. On the Rails side, when a
DnsRecord is created or updated, chuck it in the cache so that
`dns_server` can read it right away. A future nicety might be adding an
`enabled` boolean to DnsRecord to give us finer-grained control over
adding records to the cache.