Put GoodJob batches to work fetching blocklists
Ultimately, what I wanted was to fire off a batch of jobs that once
complete, could fire off a Turbo update to the UI so that I can show the
state of fetching/parsing blocklists to users. To that end, use a
GoodJob batch when parsing and creating DnsRecords from a blocklist. The
batch API is very similar to Sidekiq's. One bit of performance though,
is that a single job to create each of +100k DnsRecords is SLOOOOOOW. It
took a while with Sidekiq, too (seemed faster than GoodJob, though). To
mitigate the slowness, use the new `upsert_all` that we got in Rails 7.
Now, 150k DnsRecords can be upserted with 15 jobs, and performance is
wicked fast, seconds vs. minutes.