Start running dns_server locally for everything
Does what it says. I'm turning it on and running it for all DNS queries
on my laptop. This will allow me to gather a bunch of stats and data to
work with. Add a bunch of debugging to help me sort out what to do with
non-IN::A queries.